+++ poriscope/plugins/analysistabs/EventAnalysisView.py
@@ -39,13 +39,13 @@ from PySide6.QtWidgets import QBoxLayout, QFileDialog, QHBoxLayout, QMessageBox
 from poriscope.plugins.analysistabs.utils.eventAnalysisControls import (
     EventAnalysisControls,
 )
-from poriscope.plugins.analysistabs.utils.walkthrough_mixin import (
-    WalkthroughMixin,
-    WalkthroughStep,
-)
 from poriscope.utils.DocstringDecorator import inherit_docstrings
 from poriscope.utils.LogDecorator import log
 from poriscope.utils.MetaEventTabView import MetaEventTabView
+from poriscope.views.widgets.walkthrough_mixin import (
+    WalkthroughMixin,
+    WalkthroughStep,
+)
 
 
 @inherit_docstrings